Fix changing other services than the user expects (bsc#1165388, bsc#1174615)
As reported:

> When I select for example auditd and change its Start Mode from "On
> Boot" to "(*) Manually" (and/or its State from "Active (Running)" to
> "(*) Inactive"), the item YaST2-Second-Stage (that is two lines above
> auditd) is visually marked as to be changed instead. But the change(s)
> will be actually correctly applied to auditd.

Cause:

There is a long-standing bug in libyui-ncurses that is mostly hidden by
circumstances. This commit does not fix that bug, only works around it.

Services Manager does something other YaST modules don't do: it changes
table contents cell by cell instead of wholesale.

The ncurses (TUI) version of the table widget has a bug where this
individual cell change ignores table sorting that may be set up.

The bug seems to be old but it was masked by sorting being initially
disabled for all tables. You had to press Ctrl-O to reorder the table.
Since SLE/Leap 15.2 tables are initially sorted so the bug triggers. The
workaround is to add the `:keepSorting` option which disables both the
initial sorting and the user-triggered one.

Another factor masking the bug is the fact that application code, aware
that tables will not sort, was sorting the items before inserting them
in tables. So in many cases the application order and the table order
are the same and the bug is hidden. But if there are items in lowercase
as well as Uppercase the two sort algorithms disagree and the bug
becomes apparent.